style changes

+++ kernel/arch/i386/vga.h
@@ -24,12 +24,12 @@ enum vga_color {
 
 static inline uint8_t vga_entry_color( enum vga_color fg, enum vga_color bg ) 
 {
-	return fg | bg << 4;
+	return (uint8_t)(fg | bg << 4);
 }
 
 static inline uint16_t vga_entry( unsigned char uc, uint8_t color ) 
 {
-	return (uint16_t) uc | (uint16_t) color << 8;
+	return (uint16_t)((uint16_t) uc | (uint16_t) color << 8);
 }
 
 #endif
