+++ kernel/kernel/cpu/idt.c
@@ -32,7 +32,7 @@ void idt_load( uintptr_t );
 void idt_set_gate( uint8_t num, idt_gate_t base, uint16_t sel, uint8_t flags )
 {
     ENTRY(num).base_low = ((uintptr_t)base & 0xFFFF);
-    ENTRY(num).base_high = ((uintptr_t)base >> 16 ) & 0xFFFF;
+    ENTRY(num).base_high = (uint16_t)((uintptr_t)base >> 16 ) & 0xFFFF;
     ENTRY(num).sel = sel;
     ENTRY(num).zero = 0;
     ENTRY(num).flags = flags | 0x60;
